mcp-examples/requirements.txt
fastmcp>=1.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
logging>=0.5.0 